# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ecommerce API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://laravel-12.test"
        },
        {
            "url": "http://localhost:8000/api"
        },
        {
            "url": "http://localhost:8000/api/v1"
        },
        {
            "url": "http://localhost:8000/api/v2"
        }
    ],
    "paths": {
...
        "/api/products/{id}/inventory": {
            "patch": {
                "tags": [
                    "Inventory Management"
                ],
                "summary": "Update product inventory",
                "description": "Update inventory quantity for a specific product using set, increment, or decrement operations",
                "operationId": "0d594c792c365b1d61b11af8178788db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quantity",
                                    "operation"
                                ],
                                "properties": {
                                    "quantity": {
                                        "description": "Quantity to set, increment, or decrement",
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "operation": {
                                        "description": "Inventory operation type",
                                        "type": "string",
                                        "enum": [
                                            "set",
                                            "increment",
                                            "decrement"
                                        ],
                                        "example": "set"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product inventory updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product inventory updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Insufficient inventory",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Insufficient inventory"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to update inventory"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "title": "Product",
                "description": "Product model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "T-Shirt"
                    },
                    "description": {
                        "type": "string",
                        "example": "A comfortable cotton t-shirt"
                    },
                    "sku": {
                        "type": "string",
                        "example": "TSH-001"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 29.99
                    },
                    "compare_price": {
                        "type": "number",
                        "format": "float",
                        "example": 39.99
                    },
                    "inventory_quantity": {
                        "type": "integer",
                        "example": 100
                    },
                    "track_inventory": {
                        "type": "boolean",
                        "example": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive",
                            "archived"
                        ],
                        "example": "active"
                    },
                    "vendor": {
                        "type": "string",
                        "example": "Brand Name"
                    },
                    "product_type": {
                        "type": "string",
                        "example": "Clothing"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "cotton",
                            "casual"
                        ]
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "image1.jpg",
                            "image2.jpg"
                        ]
                    },
                    "weight": {
                        "type": "number",
                        "format": "float",
                        "example": 0.3
                    },
                    "weight_unit": {
                        "type": "string",
                        "enum": [
                            "kg",
                            "g",
                            "lb",
                            "oz"
                        ],
                        "example": "kg"
                    },
                    "requires_shipping": {
                        "type": "boolean",
                        "example": true
                    },
                    "seo": {
                        "properties": {
                            "title": {
                                "type": "string",
                                "example": "Comfortable Cotton T-Shirt"
                            },
                            "description": {
                                "type": "string",
                                "example": "Meta description for SEO"
                            },
                            "keywords": {
                                "type": "string",
                                "example": "t-shirt, cotton, clothing"
                            }
                        },
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "variants": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductVariant"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}